feat: add grid system (container) to pages
and also make featured titles on index page look nice
