+++ moviefrontend/src/App.css
@@ -36,3 +36,36 @@
     transform: rotate(360deg);
   }
 }
+
+.featured-titles-carousel {
+  max-width: 100%;
+  margin: auto;
+  height: 550px;
+}
+
+.featured-title-carousel-card {
+  object-fit: cover;
+  max-width: 80%;
+  height: 30rem;
+  display: block;
+  margin: 0 auto;
+  width: 18rem;
+}
+
+.featured-title-carousel-title {
+  font-size: 1.5rem;
+  font-weight: bold;
+  margin: 0;
+  padding: 0;
+  color: white;
+  box-shadow: 10px 5px 5px black;
+  background-color: black;
+}
+
+.title-card-image {
+  object-fit: scale-down;
+  max-width: 100%;
+  max-height: 85%;
+  display: block;
+  margin: 0 auto;
+}
